# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Users",
        "operationId": "get_users",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "type": "array",
                  "title": "Response Get Users Users Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "description": "Creates a new participant. Requires an admin's user_id for authentication.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      }
    },
...
    "/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User",
        "operationId": "get_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid4"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/responses": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User Responses",
        "operationId": "get_user_responses",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response Get User Responses Users  User Id  Responses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/tasks": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get My Tasks",
        "operationId": "get_my_tasks",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskReadParticipant"
                  },
                  "type": "array",
                  "title": "Response Get My Tasks Users Me Tasks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      }
    },
    "/tasks/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Tasks",
        "operationId": "get_tasks",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "type": "array",
                  "title": "Response Get Tasks Tasks  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task",
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Delete Task",
        "operationId": "delete_task",
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tasks/{task_id}/responses": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Responses",
        "operationId": "get_task_responses",
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponseRead"
                  },
                  "title": "Response Get Task Responses Tasks  Task Id  Responses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/response": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task Response",
        "operationId": "create_task_response",
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskResponseCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Task already completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/participants": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Participants",
        "operationId": "get_task_participants",
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskParticipantRead"
                  },
                  "title": "Response Get Task Participants Tasks  Task Id  Participants Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/responses": {
      "get": {
        "tags": [
          "responses"
        ],
        "summary": "Get Responses",
        "operationId": "get_responses",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskResponseRead"
                  },
                  "type": "array",
                  "title": "Response Get Responses Responses Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalHTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RootModel_int_": {
        "type": "integer",
        "title": "RootModel[int]"
      },
      "RootModel_list_int__": {
        "items": {
          "type": "integer"
        },
        "type": "array",
        "title": "RootModel[list[int]]"
      },
      "RootModel_str_": {
        "type": "string",
        "title": "RootModel[str]"
      },
...
      "TaskConfig-Input": {
        "properties": {
          "name": {
            "$ref": "#/components/schemas/Translations"
          },
          "description": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Translations"
              },
              {
                "type": "null"
              }
            ]
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/TaskPage-Input"
            },
            "type": "array",
            "title": "Pages"
          },
          "login_required": {
            "type": "boolean",
            "title": "Login Required"
          }
        },
        "type": "object",
        "required": [
          "name",
          "pages",
          "login_required"
        ],
        "title": "TaskConfig"
      },
      "TaskConfig-Output": {
        "properties": {
          "name": {
            "$ref": "#/components/schemas/Translations"
          },
          "description": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Translations"
              },
              {
                "type": "null"
              }
            ]
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/TaskPage-Output"
            },
            "type": "array",
            "title": "Pages"
          },
          "login_required": {
            "type": "boolean",
            "title": "Login Required"
          }
        },
        "type": "object",
        "required": [
          "name",
          "pages",
          "login_required"
        ],
        "title": "TaskConfig"
      },
      "TaskCreate": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/TaskConfig-Input"
          }
        },
        "type": "object",
        "required": [
          "config"
        ],
        "title": "TaskCreate"
      },
...
      "TaskPage-Output": {
        "properties": {
          "columns": {
            "type": "integer",
            "title": "Columns",
            "description": "The number of columns used to display the component groups",
            "default": 1
          },
          "label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Translations"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "component_groups": {
            "items": {
              "$ref": "#/components/schemas/ComponentGroup-Output"
            },
            "type": "array",
            "title": "Component Groups"
          }
        },
        "type": "object",
        "required": [
          "component_groups"
        ],
        "title": "TaskPage"
      },
      "TaskParticipantRead": {
        "properties": {
          "task": {
            "type": "string",
            "format": "uuid4",
            "title": "Task"
          },
          "user": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          }
        },
        "type": "object",
        "required": [
          "task",
          "user",
          "completed"
        ],
        "title": "TaskParticipantRead"
      },
      "TaskRead": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/TaskConfig-Output"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "config"
        ],
        "title": "TaskRead"
      },
      "TaskReadParticipant": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/TaskConfig-Output"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          }
        },
        "type": "object",
        "required": [
          "config",
          "completed"
        ],
        "title": "TaskReadParticipant"
      },
      "TaskResponse": {
        "properties": {
          "response": {
            "additionalProperties": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RootModel_int_"
                },
                {
                  "$ref": "#/components/schemas/RootModel_list_int__"
                },
                {
                  "$ref": "#/components/schemas/RootModel_str_"
                }
              ]
            },
            "type": "object",
            "title": "Response"
          },
          "updated_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Timestamp"
          },
          "created_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Created Timestamp"
          },
          "task": {
            "type": "string",
            "format": "uuid4",
            "title": "Task"
          },
          "user": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "type": "object",
        "required": [
          "response",
          "task",
          "user"
        ],
        "title": "TaskResponse"
      },
      "TaskResponseCreate": {
        "properties": {
          "response": {
            "additionalProperties": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RootModel_int_"
                },
                {
                  "$ref": "#/components/schemas/RootModel_list_int__"
                },
                {
                  "$ref": "#/components/schemas/RootModel_str_"
                }
              ]
            },
            "type": "object",
            "title": "Response"
          }
        },
        "type": "object",
        "required": [
          "response"
        ],
        "title": "TaskResponseCreate"
      },
      "TaskResponseRead": {
        "properties": {
          "response": {
            "additionalProperties": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RootModel_int_"
                },
                {
                  "$ref": "#/components/schemas/RootModel_list_int__"
                },
                {
                  "$ref": "#/components/schemas/RootModel_str_"
                }
              ]
            },
            "type": "object",
            "title": "Response"
          },
          "updated_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Timestamp"
          },
          "created_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Created Timestamp"
          },
          "task": {
            "type": "string",
            "format": "uuid4",
            "title": "Task"
          },
          "user": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "type": "object",
        "required": [
          "response",
          "task",
          "user"
        ],
        "title": "TaskResponseRead"
      },
      "Translations": {
        "properties": {
          "languages": {
            "patternProperties": {
              "^\\w{2}$": {
                "type": "string"
              }
            },
            "type": "object",
            "title": "Languages",
            "examples": [
              {
                "en": "This is the english translation.",
                "zh": "这是中文翻译。"
              }
            ]
          },
          "default": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\w{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default",
            "description": "Default language to display. Must must be in `languages`. Will display an arbitrary language if left `null`.",
            "examples": [
              "en",
              "zh"
            ]
          }
        },
        "type": "object",
        "required": [
          "languages"
        ],
        "title": "Translations"
      },
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "attributes": {
            "type": "object",
            "title": "Attributes",
            "default": {}
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "User"
      },
      "UserCreate": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "attributes": {
            "type": "object",
            "title": "Attributes",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "UserCreate"
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "attributes": {
            "type": "object",
            "title": "Attributes",
            "default": {}
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "UserRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OptionalHTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}